Consolidate test vectors, add nextest filter
