+++ crates/web5/src/lib.rs
@@ -8,3 +8,5 @@ pub mod rfc3339;
 
 #[cfg(test)]
 mod test_helpers;
+#[cfg(test)]
+mod test_vectors;
